pyproject.toml
[tool.poetry]
name = "alsalam"
version = "0.1.0"
description = ""
authors = ["Anas Hagras <anas.hagras1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
black = "^24.3.0"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
pip = "^24.0"
djangorestframework = "^3.15.1"
django-rest-knox = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
